# contracts/api.yaml
...
openapi: 3.0.3
info:
  title: "AML Case Management API"
  version: "1.0.0"
  description: |
    API contracts for the AML Case Management System.

    ## Overview

    This API enables Spriggy to meet AUSTRAC AML/CTF compliance obligations by providing
    endpoints for managing KYC remediation, PEP/Sanctions screening disposition, and
    suspicious activity investigation workflows.

    ## Authentication

    All endpoints require Bearer token authentication via OIDC SSO integration (D5).
    Include the token in the Authorization header:

    ```
    Authorization: Bearer <access_token>
    ```

    Webhook endpoints use HMAC signature validation instead of Bearer tokens (D11).

    ## Roles

    The system enforces four distinct roles (FR-063):
    - **L1_ANALYST**: Triage and basic case handling
    - **L2_ANALYST**: Full investigation and SMR recommendation authority
    - **AML_MANAGER**: Approval authority for SMRs
    - **READ_ONLY**: View and export only

    ## Error Handling

    All errors follow a consistent format with machine-readable codes.
    See the ErrorResponse schema for details.

    ## Optimistic Locking

    Case mutations require a `version` field for optimistic locking (D8).
    Version mismatch returns 409 Conflict.

    ## Constitution Alignment

    - **Principle I**: All mutating endpoints create audit log entries
    - **Principle II**: Endpoints enforce role-based access
...
  contact:
    name: AML Compliance Team
...
servers:
  - url: /api/v1
...
paths:
...
  /queue/l2-review:
    get:
      summary: List L1 closures pending L2 review
      description: |
        Retrieve cases closed by L1 with satisfactory explanation that are pending L2 quality review.

        **Filter**: Cases where l2ReviewStatus = PENDING_REVIEW

        **Roles**: L2_ANALYST, AML_MANAGER

        **Source Requirements**: FR-012, FR-018, US-16
      operationId: listL2ReviewQueue
      tags:
        - Queue
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: L2 review queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: L2+ role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /queue/l2-review/{caseId}/accept:
    parameters:
      - $ref: '#/components/parameters/CaseIdParam'

    post:
      summary: Accept L1 closure decision
      description: |
        Accept an L1 closure decision as part of quality review. Case remains closed.

        **Side Effect**: Sets l2ReviewStatus to REVIEWED_ACCEPTED

        **Roles**: L2_ANALYST, AML_MANAGER

        **Source Requirements**: FR-018
      operationId: acceptL1Closure
      tags:
        - Queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptL1ClosureRequest'
      responses:
        '200':
          description: L1 closure accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: Case not in L2 review queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /queue/unassigned:
    get:
      summary: List unassigned cases
      description: |
        Retrieve unassigned cases available for claiming, ordered by SLA urgency.

        **Roles**: L1_ANALYST, L2_ANALYST, AML_MANAGER

        **Source Requirements**: D13, FR-024, FR-025
      operationId: listUnassignedQueue
      tags:
        - Queue
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: tier
          in: query
          description: Filter by analyst tier
          schema:
            $ref: '#/components/schemas/Tier'
      responses:
        '200':
          description: Unassigned case queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # --------------------------------------------------------------------------
  # Investigation Findings
  # --------------------------------------------------------------------------
...
  /reports/sla-compliance:
    get:
      summary: Get SLA compliance report
      description: |
        Retrieve SLA compliance metrics by case type and time period.

        **Data Source**: Materialized view on read replica (D7).

        **Roles**: All authenticated users

        **Source Requirements**: FR-070
      operationId: getSLAComplianceReport
      tags:
        - Reports
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: SLA compliance report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SLAComplianceReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /reports/aged-cases:
    get:
      summary: Get aged cases report
      description: |
        Retrieve cases that have exceeded normal processing time (FR-073).

        **Data Source**: Materialized view on read replica (D7).

        **Roles**: All authenticated users

        **Source Requirements**: FR-073
      operationId: getAgedCasesReport
      tags:
        - Reports
      parameters:
        - name: ageThresholdDays
          in: query
          description: Minimum age in days to include
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Aged cases report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgedCasesReportResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /reports/export:
    post:
      summary: Export report data
      description: |
        Export report data in CSV or Excel format for governance presentation.

        **Roles**: All authenticated users

        **Source Requirements**: FR-072
      operationId: exportReport
      tags:
        - Reports
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportReportRequest'
      responses:
        '200':
          description: Exported report file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # --------------------------------------------------------------------------
  # Case Timeline and Audit
  # --------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from OIDC SSO

  # --------------------------------------------------------------------------
  # Common Parameters
  # --------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CaseIdParam:
      name: caseId
      in: path
      required: true
      description: Case unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: One or more fields are invalid
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: You do not have permission to access this resource
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: The requested resource was not found
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable message
          example: One or more fields are invalid
        details:
          type: object
          additionalProperties: true
          description: Additional error context
...
    CaseType:
      type: string
      enum:
        - KYC_REMEDIATION
        - PEP_SANCTIONS_SCREENING
        - SUSPICIOUS_ACTIVITY
        - SUPPLEMENTARY_SMR
    CaseSubtype:
      type: string
      enum:
        - SANCTIONS_ONBOARDING
        - SANCTIONS_EXISTING_CUSTOMER
        - PEP_EDD_REQUIRED
        - PEP_PROVISIONAL_REVIEW
    CaseStatus:
      type: string
      enum:
        - OPEN
        - IN_PROGRESS
        - PENDING_INFORMATION
        - ESCALATED
        - PENDING_APPROVAL
        - CLOSED
        - SMR_FILED
    L2ReviewStatus:
      type: string
      enum:
        - NOT_REQUIRED
        - PENDING_REVIEW
        - REVIEWED_ACCEPTED
        - REVIEWED_REOPENED
    Tier:
      type: string
      enum:
        - L1
        - L2
    UserRole:
      type: string
      enum:
        - L1_ANALYST
        - L2_ANALYST
        - AML_MANAGER
        - READ_ONLY
    AlertSource:
      type: string
      enum:
        - GREENID
        - INDUE
...
    CaseSummary:
      type: object
      required:
        - id
        - referenceCode
        - caseType
        - status
        - tier
        - slaDeadline
        - slaStatus
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        referenceCode:
          type: string
          example: AML-1001
        caseType:
          $ref: '#/components/schemas/CaseType'
        caseSubtype:
          $ref: '#/components/schemas/CaseSubtype'
        status:
          $ref: '#/components/schemas/CaseStatus'
        tier:
          $ref: '#/components/schemas/Tier'
        slaDeadline:
          type: string
          format: date-time
        slaStatus:
          type: string
          enum: [on_track, warning, breached]
        slaBreached:
          type: boolean
        l2ReviewStatus:
          $ref: '#/components/schemas/L2ReviewStatus'
        assignedUser:
          $ref: '#/components/schemas/UserSummary'
        customer:
          $ref: '#/components/schemas/CustomerSummary'
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    CaseDetailResponse:
      allOf:
        - $ref: '#/components/schemas/CaseSummary'
        - type: object
          required:
            - version
          properties:
            referenceNumber:
              type: integer
            source:
              $ref: '#/components/schemas/AlertSource'
            sourceAlertId:
              type: string
            pepConfidenceScore:
              type: number
              format: float
            hasCombinedAlert:
              type: boolean
            version:
              type: integer
              description: Optimistic lock version
            updatedAt:
              type: string
              format: date-time
            linkedCases:
              type: array
              items:
                $ref: '#/components/schemas/LinkedCaseSummary'
            hasOnboardingBlock:
              type: boolean
            hasEDDChecklist:
              type: boolean
            hasSMRRecommendation:
              type: boolean
            hasInvestigationFindings:
              type: boolean
    CaseListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CaseSummary'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    LinkedCaseSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        referenceCode:
          type: string
        linkType:
          type: string
          enum: [SUPPLEMENTARY_TO_ORIGINAL, RELATED, SAME_CUSTOMER]
        direction:
          type: string
          enum: [source, target]
...
    CustomerSummary:
      type: object
      description: |
        Customer snapshot captured at case creation time.

        **Immutability**: Customer data is captured as a snapshot when the case is created.
        Subsequent changes to the customer's profile in Spriggy do not update this snapshot.
        This ensures case context reflects the state at the time of the alert, which is
        important for audit and investigation purposes. If updated customer information
        is needed, analysts can reference Spriggy's customer management system directly.
      required:
        - id
        - externalId
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          description: Spriggy customer identifier
        displayName:
          type: string
          description: Masked or authorized display name
        accountStatus:
          type: string
          enum: [ACTIVE, SUSPENDED, CLOSED]
          description: Customer account status at time of case creation (snapshot)
        onboardingStatus:
          type: string
          enum: [IN_PROGRESS, COMPLETED, BLOCKED, UNKNOWN]
          description: Customer onboarding status at time of case creation (snapshot)
        snapshotAt:
          type: string
          format: date-time
          description: Timestamp when this customer data snapshot was captured

    # --- User Schemas ---
    UserSummary:
      type: object
      required:
        - id
        - displayName
        - role
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
...
    AcceptL1ClosureRequest:
      type: object
      properties:
        reviewNotes:
          type: string
          description: Optional review notes

    # --- Dashboard Schemas ---
...
    SLAComplianceReportResponse:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalCases:
              type: integer
            slaMetCount:
              type: integer
            slaBreachedCount:
              type: integer
            complianceRate:
              type: number
              format: float
        byCaseType:
          type: array
          items:
            type: object
            properties:
              caseType:
                $ref: '#/components/schemas/CaseType'
              totalCases:
                type: integer
              slaMet:
                type: integer
              slaBreached:
                type: integer
              avgResolutionDays:
                type: number
...
    AgedCasesReportResponse:
      type: object
      properties:
        summary:
          type: object
          properties:
            totalAgedCases:
              type: integer
            oldestCaseDays:
              type: integer
        cases:
          type: array
          items:
            type: object
            properties:
              caseId:
                type: string
                format: uuid
              referenceCode:
                type: string
              caseType:
                $ref: '#/components/schemas/CaseType'
              status:
                $ref: '#/components/schemas/CaseStatus'
              ageDays:
                type: integer
              assignedUser:
                $ref: '#/components/schemas/UserSummary'
    ExportReportRequest:
      type: object
      required:
        - reportType
        - format
        - startDate
        - endDate
      properties:
        reportType:
          type: string
          enum: [case_volumes, sla_compliance, smr_filings, aged_cases]
        format:
          type: string
          enum: [csv, xlsx]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    # --- Timeline and Audit Schemas ---
...
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
        hasNext:
          type: boolean
        hasPrev:
          type: boolean

# ============================================================================
# DEFAULT SECURITY
# ============================================================================
security:
  - bearerAuth: []